+++ includes/macros.h
@@ -1,0 +1,2 @@
+#define ERR_DEFAULT	"exit\n"
+#define ERR_ARG 	"Minishell doesn't take arguments\n"
