+++ srcs/check_syntax.c
@@ -1,0 +1,14 @@
+#include "minishell.h"
+
+int	check_syntax_error
+{
+	if(ft_strcmp() && prev_token == 0);
+		// print error and give back prompt
+	if ()
+		// print error and give back prompt
+	if ()
+		// print error and give back prompt
+	return (0);
+}
+
+
